+++ ui/src/pages/ChooseCategoryPage.jsx
@@ -1,3 +1,4 @@
+import React, { useEffect, useState } from 'react';
 import Card from "../components/Cards/Card";
 import CardsContainer from "../components/Cards/CardsContainer";
 import ProgressBar from "../components/ProgressBar/ProgressBar";
@@ -11,62 +12,59 @@ export default function ChooseCategoryPage() {
     const queryParams = new URLSearchParams(location.search);
     const categoryType = queryParams.get('type');
 
+    const [categories, setCategories] = useState([]);
+
+    const getCategories = async (category_type) => {
+
+        try {
+            const url = new URL(`${import.meta.env.VITE_API_ENDPOINT}/categories`);
+        
+            if (category_type) {
+                url.searchParams.append('category_type', category_type);  // Add category_type as a query parameter
+            }
+
+            const response = await fetch(url, {
+                method: 'GET',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            });
+
+            // Check if the response is ok (status code 200-299)
+            if (!response.ok) {
+                const errorData = await response.json();
+                throw new Error(errorData.detail);
+            } else {
+                // If the user was successfully created
+                const responseData = await response.json();
+                return responseData;
+            }
+        } catch (error) {
+            // Handle any errors
+            console.log(error.message);
+            return [];
+        }
+    };
+
+    useEffect(() => {
+        const fetchCategoriesData = async () => {
+            const fetchedCategories = await getCategories(categoryType)
+
+            setCategories(fetchedCategories); 
+        };
+
+        fetchCategoriesData();
+    }, [categoryType]);
+
     const handleClickCategory = (category) => {
         navigate(`/move?type=${categoryType}&category=${category}`);
     };
 
     // Need a function to break up the text
-    const handsData = [
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/sitting_alone/sitting_alone.png",
-            textSection: <p>Lead's left &lt;-&gt; follow's right</p>,
-            selectVal: "lead-left-follow-right",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/sigh/sigh.png", 
-            textSection: <p>Lead's left &lt;-&gt; follow's left</p>,
-            selectVal: "lead-left-follow-left",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/holding_back_tears/holding_back_tears.png", 
-            textSection: <p>Lead's right &lt;-&gt; follow's left</p>,
-            selectVal: "lead-right-follow-left",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/sobbing/sobbing.png", 
-            textSection: <p>Lead's right &lt;-&gt; follow's right</p>,
-            selectVal: "lead-right-follow-right",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/crying_in_hands/crying_in_hands.png", 
-            textSection: <p>Both hands straight</p>,
-            selectVal: "both-hands-straight",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/emotional_eating/emotional_eating.png", 
-            textSection: <p>Crossed hands (lead's right on top)</p>,
-            selectVal: "crossed-hands-lead-right-top",
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/angry/angry.png", 
-            textSection: <p>Crossed hands (lead's left on top)</p>,
-            selectVal: "crossed-hands-lead-left-top",
-        },
-        
-    ];
-
-    const positionsData = [
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/shame/shame.png",
-            textSection: <p>Cross Body Lead</p>,
-        },
-        {
-            image: "https://qotoqot.com/sad-animations/img/100/sleepy/sleepy.png", 
-            textSection: <p>Hammerlock</p>,
-        },        
-    ];
-
-    const cardData = categoryType === 'hands' ? handsData : positionsData;
+    const category_array = categories.map(category => ({
+        ...category,
+        image: "https://qotoqot.com/sad-animations/img/100/sitting_alone/sitting_alone.png"
+    }))
 
     return (
         <div className="main-container">
@@ -79,12 +77,12 @@ export default function ChooseCategoryPage() {
                 <CardsContainer 
                     cards={
                         <>
-                            {cardData.map((data, index) => (
+                            {category_array.map((data, index) => (
                                 <Card 
                                     key={index} 
                                     image={data.image} 
-                                    textSection={data.textSection} 
-                                    onClick={() => handleClickCategory(data.selectVal)}
+                                    textSection={data.category_name} 
+                                    onClick={() => handleClickCategory(data.category_id)}
                                 />
                             ))}
                         </>